move return_annotation of container
